requirements.txt
apache-airflow==3.1.3
dbt-duckdb==1.10.0
duckdb==1.4.2
pandas==2.3.3
pyarrow==22.0.0
black==25.11.0
pylint==4.0.4
isort==7.0.0
jq==1.10.0